requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
bcrypt==4.1.2
pyjwt==2.8.0
